This site seems to use " / " as new line delimiter
This site seems to use " / " as new line delimiter for some songs at least. The plugin ends up joining lines because of this to predictably unreadable effect.

Example: http://letras.mus.br/ryan-adams/97757/

"jesse's spinning the tuneseyelashes"

The line that ends with tunes was joined up with the word that starts the next sentence (eyelashes).

My "fix" doesn't solve the issue for some reason but I couldn't find a way of commenting or sending feedback so I'm doing it via this commit.